BuildSystem: Ensure version is udpated on change
Previously the correct version was used on a clean build, but not on
subsequent builds - the old version was used.
The manfiest now depends on a version "tag" file, that has a unique
nameper version. This ensures that the manifest is recreated (and
subsequently the jar re-packaged) as soon as the version changes, as the
appropirate version "tag" file won't exist.
(This works even for going to past revisions, as upon creating a new
version "tag", all previous ones are removed)