Fetcher: Use the field number and not index
When overriding the priting of certain fields, we were checking the
index (ie position of the field) against our map of overrides, instead
of it's unique key (which is called the number in the Java API).
This results in incorrect printing of certain fields if they are not
sent exactly in the order we expect them to (which appears to be the
case using nested messages).