Fetcher: Make it possible to clear ID fields
When using edit-config, it wasn't possible to unset an ID that was set
(either AVR of Power Board), as if the relevant option was ommited, the
value set in the old config of the node would remain.

Added a magic canary value (0) that clears the ID of the old config when
used.