WIP: Fetcher: Support for overriding system DNS
In order to avoid having to register the nodes in a DNS server,
and having to edit /etc/hosts everywhere,
we would like to allow providing hosts like file that would be used
instead.

This sets up the initial framework for doing so.
It uses a sketchy api (sun.net.spi.nameservice),
but it appears to be the only way of currently overriding the System DNS
in Java.

- Added a new javac option in the build system to supress the proprietary
  API warning.

- Added stub methods for the NameService

- Enabled the NameService in main

- Added proper magic in the META-INF/ folder,
  allowing the NameService to be loaded