Fetcher: Rework DNS overriding support
Completely rework the support for overriding the system DNS to:
    - No longer use any internal APIs
    - Not change the way DNS handled but for nodes
    - Generally be less sketchy

A new NodeAddress class now wraps InetAddress to:
    - Provide gurantees about when DNS requets will be made
    - Allow checking whether or not a host name is known
    - Use the NodeNameService with higher priority than the system DNS
      for node name resolution

The NodeOperation.processNode has been refactored to use the
NodeAddresses instead of InetAddresses.

The sketchy foo in the makefile to supress "Internal / Proprietary API"
warnings from javac has been removed.

Node addresses are now resolved after the command line arguments are
parsed, as we need to have parsed the arguments to obtain the name of
the extra hosts file to parse, which we need in turn for resolving node
addresses.

Added the commons-validator library to validate, IPv4 and IPv6
addresses, and hostnames. (The only builtin ones are internal Sun APIs).