Fetcher: Better hosts parsing error reporting
NodeNameService.parse now accepts a lambda (well.. as close as one can
get to a lambda in Java 7) that it will call to report what line number
it's parsing.
This allows the callee to set the context for logging appropriately.

The context for hgosts file parsing is now set to the "traditional"
hostfilename:linenumber.

(And yes, this is all a bit horrific, but it makes the log messages
rather nice / clear. Which will be good when the parser breaks.)